+++ link/src/test/java/com/stripe/android/link/LinkActivityContractTest.kt
@@ -4,10 +4,14 @@ import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
 import com.stripe.android.PaymentConfiguration
 import com.stripe.android.link.model.StripeIntentFixtures
+import com.stripe.android.networking.StripeRepository
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.Mockito.mock
+import org.mockito.kotlin.any
+import org.mockito.kotlin.whenever
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
@@ -43,7 +47,9 @@ class LinkActivityContractTest {
             config,
             null,
         )
-        val contract = LinkActivityContract()
+        val stripeRepository = mock<StripeRepository>()
+        whenever(stripeRepository.buildPaymentUserAgent(any())).thenReturn("test")
+        val contract = LinkActivityContract(stripeRepository)
         val intent = contract.createIntent(ApplicationProvider.getApplicationContext(), args)
         assertThat(intent.component?.className).isEqualTo(LinkForegroundActivity::class.java.name)
         assertThat(intent.extras?.getString(LinkForegroundActivity.EXTRA_POPUP_URL)).startsWith(