+++ paymentsheet/src/main/java/com/stripe/android/customersheet/injection/CustomerSessionComponent.kt
@@ -7,12 +7,18 @@ import com.stripe.android.customersheet.CustomerSheet
 import com.stripe.android.customersheet.CustomerSheetResultCallback
 import com.stripe.android.customersheet.CustomerSheetViewModel
 import com.stripe.android.customersheet.ExperimentalCustomerSheetApi
+import com.stripe.android.payments.core.injection.StripeRepositoryModule
 import dagger.BindsInstance
 import dagger.Component
 
 @OptIn(ExperimentalCustomerSheetApi::class)
 @CustomerSessionScope
-@Component(modules = [CustomerSheetViewModelModule::class])
+@Component(
+    modules = [
+        CustomerSheetViewModelModule::class,
+        StripeRepositoryModule::class,
+    ]
+)
 internal interface CustomerSessionComponent {
     val customerSheetComponentBuilder: CustomerSheetComponent.Builder